GRN: open the add-crop form in a modal from the designer's bed inspector
The designer's "+ Add crop" link used to navigate away to /crops/new,
which lost the user's place in the canvas. Now it opens the same form
in a modal so the bed context stays visible.

Implementation
- Extracts CropForm out of NewCropPage so it's reusable. The form takes
  onSuccess + onCancel callbacks (caller decides what "done" means) and
  an optional lockedBed prop. When lockedBed is set, the bed selector
  is replaced with a "Adding to <bed name>" indicator and the bed list
  query is skipped entirely.
- NewCropPage is now a thin wrapper that reads ?bedId= from the URL,
  resolves the bed via listMyBeds, and feeds it to the form. (Bonus:
  the existing /crops/new?bedId=... links from elsewhere now actually
  pre-select the bed — they didn't before.)
- Adds a minimal Modal primitive backed by the native <dialog> element
  so we get focus trap, Esc-to-close, accessible labelling, and an
  inert background for free. Backdrop-click also closes.
- New AddCropModal mounts CropForm inside the modal with lockedBed and
  closes on success — the new crop chip appears on the bed
  immediately because CropForm already invalidates the myCrops query.
- Replaces the BedInspector Link with a button that opens the modal,
  styled the same. Disabled when the inspector is read-only.

Out of scope (mentioned but explicitly deferred): the My garden page's
"Add Crop" button still navigates to /crops/new. Modal-ising it makes
sense in a follow-up if we want consistency, but full-page makes more
sense in "manage my crops" mode than in "drop a quick crop on this
bed" mode, so leaving it for now.

Gates: typecheck, eslint, vite build, perf budget (944 KB total — main
bundle unchanged at 200 KB; the designer chunk grew ~2 KB), 236 vitest
tests.
